Amend add command
