+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -21,11 +21,11 @@ public class DeleteCommand extends Command {
 
     public static final String COMMAND_WORD = "delete";
     public static final String MESSAGE_USAGE = COMMAND_WORD
-        + ": Deletes the client identified by the client id used in the displayed client list.\n"
-        + "Parameters: "
-        + "CLIENT ID "
-        + "\n"
-        + "Example: " + COMMAND_WORD + " 1";
+            + ": Deletes the client identified by the client id used in the displayed client list.\n"
+            + "Parameters: "
+            + "CLIENT ID "
+            + "\n"
+            + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_DELETE_CLIENT_SUCCESS = "Deleted Client: %1$s";
     private final List<ClientId> clientIds;
@@ -54,7 +54,7 @@ public class DeleteCommand extends Command {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-            || (other instanceof DeleteCommand // instanceof handles nulls
-            && clientIds.equals(((DeleteCommand) other).clientIds)); // state check
+                || (other instanceof DeleteCommand // instanceof handles nulls
+                && clientIds.equals(((DeleteCommand) other).clientIds)); // state check
     }
 }
