+++ src/main/java/seedu/address/logic/parser/EditCommandParser.java
@@ -17,9 +17,9 @@ import java.util.Set;
 import java.util.function.BiConsumer;
 
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditClientDescriptor;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
+import seedu.address.model.client.Client.EditClientDescriptor;
 import seedu.address.model.client.ClientId;
 import seedu.address.model.tag.Tag;
