+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -4,8 +4,6 @@ import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.testutil.TypicalClients.getTypicalAddressBook;
 
-import java.util.function.Function;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -13,7 +11,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.client.Client;
-import seedu.address.model.client.ClientId;
+import seedu.address.model.client.Client.EditClientDescriptor;
 import seedu.address.testutil.ClientBuilder;
 
 /**
@@ -30,11 +28,10 @@ public class AddCommandIntegrationTest {
 
     @Test
     public void execute_newClient_success() {
-        Function<ClientId, Client> validClientFunction = new ClientBuilder().buildFunction();
-        Client validClient = new ClientBuilder().build();
+        EditClientDescriptor validClientFunction = new ClientBuilder().buildFunction();
 
         Model expectedModel = new ModelManager(model.getAddressBook(), new UserPrefs());
-        expectedModel.addClient(new ClientBuilder().build());
+        Client validClient = expectedModel.createClient(validClientFunction);
 
         assertCommandSuccess(new AddCommand(validClientFunction), model,
             String.format(AddCommand.MESSAGE_SUCCESS, validClient), expectedModel);
@@ -43,7 +40,8 @@ public class AddCommandIntegrationTest {
     @Test
     public void execute_duplicateClient_throwsCommandException() {
         Client clientInList = model.getAddressBook().getClientList().get(0);
-        assertCommandFailure(new AddCommand(x -> clientInList), model, AddCommand.MESSAGE_DUPLICATE_CLIENT);
+        EditClientDescriptor editClientDescriptor = new ClientBuilder(clientInList).buildFunction();
+        assertCommandFailure(new AddCommand(editClientDescriptor), model, AddCommand.MESSAGE_DUPLICATE_CLIENT);
     }
 
 }
