+++ src/test/java/seedu/address/model/ModelManagerTest.java
@@ -78,22 +78,6 @@ public class ModelManagerTest {
     }
 
     @Test
-    public void hasClient_nullClient_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> modelManager.hasClient(null));
-    }
-
-    @Test
-    public void hasClient_clientNotInAddressBook_returnsFalse() {
-        assertFalse(modelManager.hasClient(ALICE));
-    }
-
-    @Test
-    public void hasClient_clientInAddressBook_returnsTrue() {
-        modelManager.addClient(ALICE);
-        assertTrue(modelManager.hasClient(ALICE));
-    }
-
-    @Test
     public void getFilteredClientList_modifyList_throwsUnsupportedOperationException() {
         assertThrows(UnsupportedOperationException.class, () -> modelManager.getFilteredClientList().remove(0));
     }
