+++ src/test/java/seedu/address/testutil/TypicalClients.java
@@ -96,8 +96,10 @@ public class TypicalClients {
      */
     public static AddressBook getTypicalAddressBook() {
         AddressBook ab = new AddressBook();
+        ab.setClientCounter("0");
         for (Client client : getTypicalClients()) {
             ab.addClient(client);
+            ab.incrementClientCounter();
         }
         return ab;
     }
